requirements.txt
requests
pytest
coverage
mongoengine
Unidecode
beautifulsoup4
python-telegram-bot
